pyproject.toml
[tool.poetry]
name = "artifactory"
version = "0.1.0"
description = ""
authors = ["Kolja huang <koljahuang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
argparse = "^1.4.0"
dynaconf = "^3.2.6"
tldextract = "^5.1.3"
pre-commit = "^4.2.0"
black = "^25.1.0"
autopep8 = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
